% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors}
\alias{colors}
\title{Color function}
\usage{
colors(yellow, red, blue)
}
\arguments{
\item{yellow}{- shade of yellow (pixels)}

\item{red}{- shade of red (pixels)}

\item{blue}{shade of blue (pixels)}
}
\description{
Color function
}
