% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snacks.R
\name{hi}
\alias{hi}
\title{function}
\usage{
hi(people, places, things, time)
}
\arguments{
\item{people}{number of people (integer)}

\item{places}{number of places (integer)}

\item{things}{number of things (cont)}

\item{time}{amount of time (seconds)}
}
\description{
function
}
\examples{

hi(people = 5, places = 4, things = 1, time = 0.5)
}
