% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snacks.R
\name{snacks}
\alias{snacks}
\title{calculate my snacks}
\usage{
snacks(grapes, apples, burritos)
}
\arguments{
\item{grapes}{mass of grapes (g)}

\item{apples}{number of apples in a bucket}

\item{burritos}{how many burritos i eat}
}
\description{
calculate my snacks
}
\examples{
snacks(grapes = 2.5, apples = 6, burritos = 1)
}
